# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart Classroom Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /classes:
    get:
      operationId: getClasses
      tags: [classes]
      summary: Get all classes
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClassWithAllocations"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a new class (HOD only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassRequest"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassWithAllocations"
  /classes/{id}:
    get:
      operationId: getClass
      tags: [classes]
      summary: Get class details
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Class details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassWithAllocations"
  /classes/{id}/allocations:
    post:
      operationId: createAllocation
      tags: [classes]
      summary: Allocate staff to a subject in a class (HOD only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAllocationRequest"
      responses:
        "201":
          description: Allocation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Allocation"
...
  /attendance/my:
    get:
      operationId: getMyAttendance
      tags: [attendance]
      summary: Get current student's attendance history
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Attendance history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
  /attendance/class/{classId}:
    get:
      operationId: getClassAttendance
      tags: [attendance]
      summary: Get attendance for a class (Staff/HOD only)
      security:
        - bearerAuth: []
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Class attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
...
  /availability:
    post:
      operationId: markUnavailable
      tags: [availability]
      summary: Mark staff as unavailable
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkUnavailableRequest"
      responses:
        "201":
          description: Availability marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Availability"
    get:
      operationId: getAvailability
      tags: [availability]
      summary: Get availability records
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Availability records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Availability"
  /burnout/staff/{staffId}:
    get:
      operationId: getStaffBurnout
      tags: [burnout]
      summary: Get burnout score for a staff member
      security:
        - bearerAuth: []
      parameters:
        - name: staffId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Burnout analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BurnoutAnalysis"
  /burnout/all:
    get:
      operationId: getAllBurnout
      tags: [burnout]
      summary: Get burnout scores for all staff (HOD only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All staff burnout data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffBurnoutSummary"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [STAFF, HOD, STUDENT]
        department:
          type: string
        phone:
          type: string
        designation:
          type: string
      required:
        - email
        - password
        - name
        - role
        - department
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [STAFF, HOD, STUDENT]
        department:
          type: string
        phone:
          type: string
        staffId:
          type: string
        studentId:
          type: string
      required:
        - id
        - email
        - name
        - role
        - department
...
    ClassWithAllocations:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        department:
          type: string
        semester:
          type: integer
        studentCount:
          type: integer
        allocations:
          type: array
          items:
            $ref: "#/components/schemas/AllocationDetail"
        studentEnrollment:
          type: integer
      required:
        - id
        - name
        - department
        - semester
        - studentCount
    CreateClassRequest:
      type: object
      properties:
        name:
          type: string
        department:
          type: string
        semester:
          type: integer
        studentCount:
          type: integer
      required:
        - name
        - department
        - semester
        - studentCount
    Allocation:
      type: object
      properties:
        id:
          type: string
        classId:
          type: string
        staffId:
          type: string
        subjectId:
          type: string
        type:
          type: string
          enum: [THEORY, LAB]
      required:
        - id
        - classId
        - staffId
        - subjectId
        - type
    AllocationDetail:
      type: object
      properties:
        id:
          type: string
        classId:
          type: string
        staffId:
          type: string
        staffName:
          type: string
        subjectId:
          type: string
        subjectName:
          type: string
        subjectCode:
          type: string
        type:
          type: string
          enum: [THEORY, LAB]
      required:
        - id
        - classId
        - staffId
        - staffName
        - subjectId
        - subjectName
        - subjectCode
        - type
    CreateAllocationRequest:
      type: object
      properties:
        staffId:
          type: string
        subjectId:
          type: string
        type:
          type: string
          enum: [THEORY, LAB]
      required:
        - staffId
        - subjectId
        - type
...
    Attendance:
      type: object
      properties:
        id:
          type: string
        studentId:
          type: string
        studentName:
          type: string
        classId:
          type: string
        className:
          type: string
        subjectName:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - studentId
        - classId
        - className
        - timestamp
...
    MarkUnavailableRequest:
      type: object
      properties:
        date:
          type: string
          format: date
        reason:
          type: string
          enum: [sick, leave, emergency, other]
        notes:
          type: string
      required:
        - date
        - reason
    Availability:
      type: object
      properties:
        id:
          type: string
        staffId:
          type: string
        staffName:
          type: string
        date:
          type: string
          format: date-time
        isAvailable:
          type: boolean
        reason:
          type: string
      required:
        - id
        - staffId
        - date
        - isAvailable
    BurnoutAnalysis:
      type: object
      properties:
        staffId:
          type: string
        staffName:
          type: string
        burnoutScore:
          type: number
        riskLevel:
          type: string
          enum: [low, medium, high]
        weeklyHours:
          type: number
        consecutiveDays:
          type: integer
        substituteCount:
          type: integer
        averageSentiment:
          type: number
        recommendations:
          type: array
          items:
            type: string
        breakdown:
          type: object
          properties:
            weeklyHoursScore:
              type: number
            consecutiveDaysScore:
              type: number
            substituteFrequencyScore:
              type: number
            sentimentScore:
              type: number
          required:
            - weeklyHoursScore
            - consecutiveDaysScore
            - substituteFrequencyScore
            - sentimentScore
      required:
        - staffId
        - staffName
        - burnoutScore
        - riskLevel
        - weeklyHours
        - consecutiveDays
        - substituteCount
    StaffBurnoutSummary:
      type: object
      properties:
        staffId:
          type: string
        staffName:
          type: string
        department:
          type: string
        burnoutScore:
          type: number
        riskLevel:
          type: string
          enum: [low, medium, high]
        weeklyHours:
          type: number
      required:
        - staffId
        - staffName
        - department
        - burnoutScore
        - riskLevel
        - weeklyHours
...